Moved error and markdown to specific files
